refactor(ui): extract AvailableCellGrid / EffectiveLabourTotal from production_panel.dart
Splits the two presentational widgets used by the Available subpanel
into `production_available_grid.dart` so `production_panel.dart` drops
back below the 700-line `repo.game_widgets_file_size` limit enforced by
the CI Quality lint (was 727 lines, now 642). Behavior is unchanged;
classes are renamed from private `_AvailableCellGrid` /
`_EffectiveLabourTotal` to public `AvailableCellGrid` /
`EffectiveLabourTotal` so they remain accessible from
`_AvailableSubpanel` in the original file, mirroring the existing
extraction pattern in this folder (e.g. `ProductionAllocationRowChrome`,
`ProductionLabourSection`).

Refs #2862
