+++ app/lib/features/game/widgets/production_panel.dart
@@ -4,7 +4,6 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:flutter/material.dart';
 
 import '../../../config/constants.dart';
-import '../../../config/editorial_monocle_palette.dart';
 import '../../../l10n/l10n.dart';
 import '../../../widgets/ct_brass_divider.dart';
 import '../../../widgets/ct_nine_patch_button.dart';
@@ -16,6 +15,7 @@ import '../../../widgets/resource_icon.dart';
 import 'chrome/ct_danger_text_button.dart';
 import 'production_allocation_row.dart';
 import 'production_allocation_row_chrome.dart';
+import 'production_available_grid.dart';
 import 'production_labour_helpers.dart';
 import 'production_labour_section.dart';
 
@@ -276,7 +276,7 @@ class _AvailableSubpanel extends StatelessWidget {
     Map<String, int> netChanges,
     Map<CommodityId, int> sellableByCommodityId,
   ) {
-    return _AvailableCellGrid(
+    return AvailableCellGrid(
       key: ValueKey<String>(
         'production_available_commodity_grid_${commodities.map((c) => c.id).join('_')}',
       ),
@@ -362,7 +362,7 @@ class _AvailableSubpanel extends StatelessWidget {
       const SizedBox(height: 12),
       CtSectionLabel(l10n.production_workers),
       const SizedBox(height: 6),
-      _AvailableCellGrid(
+      AvailableCellGrid(
         key: kProductionAvailableWorkerGridKey,
         columnCount: kProductionAvailableWorkerGridColumns,
         cells: <Widget>[
@@ -373,7 +373,7 @@ class _AvailableSubpanel extends StatelessWidget {
         ],
       ),
       const SizedBox(height: 8),
-      _EffectiveLabourTotal(
+      EffectiveLabourTotal(
         text: l10n.production_effectiveLabour(effectiveLabour),
         theme: theme,
       ),
@@ -463,91 +463,6 @@ class _AvailableSubpanel extends StatelessWidget {
           children: _buildBodyChildren(theme),
         ),
       ),
-    );
-  }
-}
-
-/// Renders a fixed-column grid of `CtResourceCell`-shaped cells for the
-/// Available subpanel (Refs #2862 S8b, owner decision **C7**).
-///
-/// Cells are laid out via `LayoutBuilder` + `Wrap` so each cell occupies a
-/// `(maxWidth - (columnCount - 1) * spacing) / columnCount` slot. The trailing
-/// row may be partially filled when `cells.length` is not a multiple of
-/// [columnCount]; the remaining slot space is left empty so the grid keeps
-/// equal column widths even when only one or two cells are present (e.g. a
-/// Manufactured section with a single recipe output).
-///
-/// SPEC: `SPEC/ui/production-panel.md` § Layout — Available subpanel
-/// "Commodity grid layout" / "Workers section".
-class _AvailableCellGrid extends StatelessWidget {
-  const _AvailableCellGrid({
-    super.key,
-    required this.columnCount,
-    required this.cells,
-  }) : assert(columnCount > 0);
-
-  final int columnCount;
-  final List<Widget> cells;
-
-  static const double _columnSpacing = 6;
-  static const double _rowSpacing = 4;
-
-  @override
-  Widget build(BuildContext context) {
-    if (cells.isEmpty) {
-      return const SizedBox.shrink();
-    }
-    return LayoutBuilder(
-      builder: (BuildContext context, BoxConstraints constraints) {
-        final double maxWidth = constraints.maxWidth.isFinite
-            ? constraints.maxWidth
-            : MediaQuery.sizeOf(context).width;
-        final double cellWidth =
-            (maxWidth - (columnCount - 1) * _columnSpacing) / columnCount;
-        final double effectiveCellWidth = cellWidth.isFinite && cellWidth > 0
-            ? cellWidth
-            : maxWidth;
-        return Wrap(
-          spacing: _columnSpacing,
-          runSpacing: _rowSpacing,
-          children: <Widget>[
-            for (final Widget cell in cells)
-              SizedBox(width: effectiveCellWidth, child: cell),
-          ],
-        );
-      },
-    );
-  }
-}
-
-/// Right-aligned **Effective labour** total rendered with the dark
-/// editorial-monocle accent color, monospace tabular figures, and a
-/// 1px `--accent-dim` top border, per the issue #2862 Available subpanel
-/// requirement R8 (right-aligned, accent color, bordered top).
-class _EffectiveLabourTotal extends StatelessWidget {
-  const _EffectiveLabourTotal({required this.text, required this.theme});
-
-  final String text;
-  final ThemeData theme;
-
-  @override
-  Widget build(BuildContext context) {
-    final base = theme.textTheme.bodySmall ?? const TextStyle(fontSize: 12);
-    final style = base.copyWith(
-      color: EditorialMonoclePalette.accent,
-      fontWeight: FontWeight.w600,
-      fontFamilyFallback: const <String>['SF Mono', 'Menlo', 'monospace'],
-      fontFeatures: const <FontFeature>[FontFeature.tabularFigures()],
-    );
-    return Container(
-      width: double.infinity,
-      decoration: BoxDecoration(
-        border: Border(
-          top: BorderSide(color: EditorialMonoclePalette.accentDim, width: 1),
-        ),
-      ),
-      padding: const EdgeInsets.only(top: 4),
-      child: Text(text, style: style, textAlign: TextAlign.right),
     );
   }
 }
